Renamed test-options to fixtures
