+++ tests/tryTrial.spec.ts
@@ -1,5 +1,5 @@
 import { faker } from '@faker-js/faker';
-import { test } from '../page-objects/test-options';
+import { test } from '../page-objects/fixtures';
 
 test.beforeEach(async ({ navigateTo }) => {
     await navigateTo('.try');
