+++ tests/video.spec.ts
@@ -1,4 +1,4 @@
-import { test } from '../page-objects/test-options';
+import { test } from '../page-objects/fixtures';
 
 test.beforeEach(async ({ navigateTo }) => {
     await navigateTo('.hero-baner#home');
